update in Supplies

+++ app/Http/Controllers/SupgenController.php
@@ -3,12 +3,18 @@
 namespace App\Http\Controllers;
 
 use App\Supgen;
-use Illuminate\Http\Request;
 use Illuminate\Support\Str;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\SupgenRequest;
 
 class SupgenController extends Controller
 {
+    public function __construct()
+    {
+        return $this->middleware('auth');
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -16,8 +22,14 @@ class SupgenController extends Controller
      */
     public function index()
     {
-        $supgens = Supgen::orderBy('id', 'desc')->get();
-        return view('inventory.supply.supply_generic.index', compact('supgens'));
+        if (Gate::check('isAdmin') || Gate::check('isDoctor') || Gate::check('isNurse') || Gate::check('isHr')) {
+            $supgens = Supgen::orderBy('id', 'desc')->paginate(10);
+            return view('inventory.supply.supply_generic.index', compact('supgens'));
+        }elseif (Gate::check('isBanned')) {
+            return back()->with('message', 'You\'re not employee!');
+        }else{
+            return back();
+        }
     }
 
     /**
@@ -38,10 +50,16 @@ class SupgenController extends Controller
      */
     public function store(SupgenRequest $request)
     {
-        $atts = $this->validate($request, $request->rules(), $request->messages());
-        $atts['slug'] = Str::slug($request->name, '-');
-        Supgen::create($atts);
-        return back();
+        if (Gate::check('isAdmin') || Gate::check('isDoctor') || Gate::check('isNurse')) {
+            $atts = $this->validate($request, $request->rules(), $request->messages());
+            $atts['slug'] = Str::slug($request->name, '-');
+            Supgen::create($atts);
+            return back();
+        }elseif (Gate::check('isBanned')) {
+            return back()->with('message', 'You\'re not employee!');
+        }else{
+            return back();
+        }
     }
 
     /**
@@ -52,7 +70,13 @@ class SupgenController extends Controller
      */
     public function show(Supgen $supgen)
     {
-        return view('inventory.supply.supply_generic.show', compact('supgen'));
+        if (Gate::check('isAdmin') || Gate::check('isDoctor') || Gate::check('isNurse') || Gate::check('isHr')) {
+            return view('inventory.supply.supply_generic.show', compact('supgen'));
+        }elseif(Gate::check('isBanned')){
+            return back()->with('message', 'You\'re not employee!');
+        }else{
+            return back();
+        }
     }
 
     /**
@@ -75,7 +99,24 @@ class SupgenController extends Controller
      */
     public function update(Request $request, Supgen $supgen)
     {
-        //
+        if (Gate::check('isAdmin') || Gate::check('isDoctor') || Gate::check('isNurse')) {
+            $atts = $request->validate(
+                [
+                    'name'  =>  ['required', 'unique:supgens,name,'.$supgen->id],
+                ],
+                [
+                    'name.required' =>  'This field is required!',
+                    'name.unique'   =>  'Name is already taken!',
+                ]
+            );
+            $atts['slug'] = Str::slug($request->name, '-');
+            $supgen->update($atts);
+            return back();
+        }elseif (Gate::check('isBanned')) {
+            return back()->with('message', 'You\'re not employee!');
+        }else{
+            return back();
+        }
     }
 
     /**
@@ -86,10 +127,16 @@ class SupgenController extends Controller
      */
     public function destroy(Supgen $supgen)
     {
-        if ($supgen->supbrands->count() < 1) {
-            $supgen->delete();
+        if (Gate::check('isAdmin') || Gate::check('isDoctor') || Gate::check('isNurse')) {
+            if ($supgen->supbrands->count() < 1) {
+                $supgen->delete();
+                return back();
+            }
+            return back()->with('supgen_error', 'You Cannot delete a Supply Name with more than 1 quantity!');
+        }elseif (Gate::check('isBanned')) {
+            return back()->with('message', 'You\'re not employee!');
+        }else{
             return back();
         }
-        return back();
     }
 }
